+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - [FIX] Group resource creation fixed.
 - [IMPROVEMENT] Now using logging for debug output.
 - [IMPROVEMENT] Simplified marshalling.
+- [IMPROVEMENT] Add Originator to Resources.
+
 
 ## Version 0.7 (2018-05-13)
 - [IMPROVEMENT] Added support for &lt;remoteCSE> resource type.